Move lua maths library to LuaMaths
