+++ Resources/Scripts/vec3_test.lua
@@ -1,0 +1,4 @@
+function Init(entity) end
+
+function OnUpdate(entity, dt)
+end
