+++ Sprocket/CMakeLists.txt
@@ -90,6 +90,7 @@ add_library(Sprocket STATIC
 
             Scripting/LuaScript.cpp
             Scripting/LuaLibrary.cpp
+            Scripting/LuaMaths.cpp
             Scripting/LuaConverter.cpp)
 
 target_link_libraries(Sprocket PUBLIC
