+++ Sprocket/Scripting/LuaMaths.h
@@ -1,0 +1,13 @@
+#pragma once
+
+struct lua_State;
+
+namespace Sprocket {
+namespace lua {
+
+class Script;
+
+void load_vec3_functions(lua::Script& script);
+
+}
+}
